// front-end/src/models/estoqueRetirada.model.ts
export class EstoqueRetirada {
    id?: number | undefined;
    item_id: number | undefined;
    marca_id: number | undefined;
    modelo: string | undefined;
    numero_serie: string | undefined;
    patrimonio: string | undefined;
    lote: string | undefined;
    quantidade: number | undefined;
    descricao: string | undefined;
}